[SON-387] Refactor serf connection to automatically close on error (#2)
[SON-387] Refactor serf connection to automatically close on error

Previous version spawned two threads, one for reading and one for
writing. These threads would call unwrap which could lead to an
undefined state since the threads were never joined upon and the default
behaviour in rust is to only unwind a thread on panic. This commit adds
an additional thread that watches the read and write tasks and on error
will close both and send errors through the dispatch handler.