Add support for the query stream (#4)
In SON-1945 we require support for query(). Query is a stream, just like
the stream call.

- Replaced example with a command line tool to execute commands just
like the `serf` tool itself. This was essential to debug serf behavior.
- Replaced `drop(stream)` with shutdown. The streams on the other
threads were not closed, causing the tokio runtime to hang on the
spawn_blocking tasks.
- Parse values into a `rmpv::Value` first. Such that we can log the
message, even if we don't know exactly which (serde) message is
incoming.
- Replaced `deserialize_payload` with `serde_bytes` directly.
- Changed payload type of incoming messages to `Option<Vec<u8>>` as
parsing would otherwise fail on empty payloads.
- Ran cargo fmt.
- Fixed various clippy lints.
- Added Clone to the exported types.

**IMPORTANT**:
[documentation](https://github.com/hashicorp/serf/blob/master/docs/agent/rpc.html.markdown#query)
of serf was lacking. Not all parameters were listed. In the actual
source code of the Go serf CLI, we can find all parameters:
https://github.com/hashicorp/serf/blob/master/client/const.go#L148